requirements.txt
pyaudio==0.2.13
google-generativeai==0.6.1
keyboard==0.13.5
pyautogui==0.9.54
pillow==10.0.0
python-dotenv==1.0.0
pystray==0.19.4
pyperclip==1.8.2 